+++ symja_android_library/matheclipse-core/src/main/java/de/tilman_neumann/jml/factor/base/congruence/Smooth_Perfect.java
@@ -34,7 +34,7 @@ public class Smooth_Perfect extends Smooth_Simple {
 	}
 	
 	@Override
-	public SortedMultiset<Integer> getAllQFactors() {
+	public SortedMultiset<Long> getAllQFactors() {
 		// a perfect smooth congruence has no large factors
 		return super.getSmallQFactors();
 	}
